+++ CommHandler/CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - client for SGr declaration library
 
+### Changed
+
+- requires updated driver implementations
+
 
 ## [2.4.2] - 2025-10-09